Use Cursor to implement fold
This fixes the FIXME about name clashes, so it should be possible to
nest 'fold' calls now.  It should also render the "FETCH FORWARD n"
query a little more efficiently since it doesn't have to search the
query string for '?'.

It does not, however, provide a better implementation for 'Automatic'
fetch sizing.